+++ module/Task/src/Domain/Task/Event/TaskDeleted.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ExtendsSoftware\ExaPHPExample\Task\Domain\Task\Event;
+
+use ExtendsSoftware\ExaPHPExample\Shared\Domain\DomainEventInterface;
+use ExtendsSoftware\ExaPHPExample\Task\Domain\Task\TaskId;
+
+final readonly class TaskDeleted implements DomainEventInterface
+{
+    public function __construct(private(set) TaskId $taskId) {}
+}